init beta version of nuxt-dapp
