--- layouts/default.vue
@@ -1,11 +1,0 @@
-<script lang="ts" setup></script>
-
-<template>
-	<div>
-		<Web3Provider>
-			<TheHeader />
-			<slot />
-			<TheFooter />
-		</Web3Provider>
-	</div>
-</template>
