+++ package.json
@@ -1,40 +1,24 @@
 {
-	"name": "nuxt-dapp",
-	"version": "1.0.0-beta.0",
-	"private": true,
-	"type": "module",
-	"scripts": {
-		"build": "nuxt build",
-		"dev": "nuxt dev",
-		"generate": "nuxt generate",
-		"preview": "nuxt preview",
-		"postinstall": "nuxt prepare"
-	},
-	"dependencies": {
-		"@pinia-plugin-persistedstate/nuxt": "^1.2.0",
-		"@pinia/nuxt": "^0.5.1",
-		"@vue-dapp/core": "1.0.0-beta.0",
-		"@vue-dapp/vd-board": "1.0.0-beta.0",
-		"@vue-dapp/walletconnect": "1.0.0-beta.0",
-		"@vueuse/core": "^10.7.1",
-		"@vueuse/nuxt": "^10.7.1",
-		"copy-to-clipboard": "^3.3.3",
-		"ethers": "^6.9.1",
-		"nuxt-icon": "^0.6.8",
-		"pinia": "^2.1.7",
-		"tiny-invariant": "^1.3.1",
-		"viem": "^1.21.1",
-		"vue3-select": "^0.1.2"
-	},
-	"devDependencies": {
-		"@nuxt/devtools": "^1.0.6",
-		"@nuxtjs/tailwindcss": "^6.10.3",
-		"autoprefixer": "^10.4.16",
-		"nuxt": "^3.9.0",
-		"postcss": "^8.4.32",
-		"postcss-custom-properties": "^13.3.3",
-		"sass": "^1.69.6",
-		"vue": "^3.4.0",
-		"vue-router": "^4.2.5"
-	}
+  "name": "nuxt-app",
+  "private": true,
+  "type": "module",
+  "scripts": {
+    "build": "nuxt build",
+    "dev": "nuxt dev",
+    "generate": "nuxt generate",
+    "preview": "nuxt preview",
+    "postinstall": "nuxt prepare"
+  },
+  "devDependencies": {
+    "nuxt": "^3.9.0",
+    "vue": "^3.4.3",
+    "vue-router": "^4.2.5"
+  },
+  "dependencies": {
+    "@pinia/nuxt": "^0.5.1",
+    "@vue-dapp/core": "1.0.0-beta.3",
+    "@vue-dapp/vd-board": "1.0.0-beta.2",
+    "@vue-dapp/walletconnect": "1.0.0-beta.2",
+    "pinia": "^2.1.7"
+  }
 }
