--- pages/index.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts"></script>
-
-<template>
-	<div></div>
-</template>
-
-<style lang="scss"></style>
