--- styles/breakpoints.json
@@ -1,7 +1,0 @@
-{
-	"sm": "480px",
-	"md": "768px",
-	"lg": "976px",
-	"xl": "1440px",
-	"2xl": "1536px"
-}
