--- styles/vars.scss
@@ -1,17 +1,0 @@
-:root {
-	--color-primary-dark: #586063;
-	--color-primary: #ababab;
-	--color-primary-light: #edecea;
-
-	--color-secondary-dark: #016665;
-	--color-secondary: #00807e;
-	--color-secondary-light: #65b1b1;
-
-	// app
-	--header-height: 48px;
-	--sidebar-width: 240px;
-
-	// vue3-select
-	--vs-dropdown-option--active-bg: var(--color-primary-dark);
-	--vs-dropdown-option--selected-bg: var(--color-primary-dark);
-}
